+++ server/asp-backend/Diplom/Models/Requests/Breakpoint.cs
@@ -5,9 +5,9 @@ namespace Diplom.Models.Requests
     public class Breakpoint
     {
         [JsonPropertyName("x")]
-        public int X { get; set; }
+        public float X { get; set; }
 
         [JsonPropertyName("y")]
-        public int Y { get; set; }
+        public float Y { get; set; }
     }
 }
